+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Unreleased
 
+* Add `ValueUtil.addSeconds(...)` utility method.
+
 ### [v0.105](https://github.com/realityforge/guiceyloops/tree/v0.105) (2019-09-25) · [Full Changelog](https://github.com/realityforge/guiceyloops/compare/v0.104...v0.105)
 
 * Upgrade the `org.realityforge.javax.annotation` artifact to version `1.0.1`.